refactor: remove http helper module
