+++ src/routes/admin/admin.ts
@@ -1,6 +1,5 @@
-import type { AppContext } from '../../type.js';
-import type { AppContextContext, AppEnv } from '../../http.js';
-import { renderView, setFlash } from '../../http.js';
+import type { AppContext, AppContextContext, AppEnv } from '../../type.js';
+import { renderView, setFlash } from '../middleware.js';
 import { Hono } from 'hono';
 
 export function createAdminRouter(ctx: AppContext) {
