+++ src/routes/general/general.ts
@@ -1,7 +1,14 @@
 import { bangs } from '../../db/bang.js';
-import type { AppContextContext, AppEnv } from '../../http.js';
-import { AppResponse, createAppRequest, renderView, setFlash } from '../../http.js';
-import type { Bang, User, AppContext, BangWithLowercase } from '../../type.js';
+import { AppResponse } from '../../type.js';
+import type {
+    AppContext,
+    AppContextContext,
+    AppEnv,
+    Bang,
+    BangWithLowercase,
+    User,
+} from '../../type.js';
+import { createAppRequest, renderView, setFlash } from '../middleware.js';
 import { Hono } from 'hono';
 
 export function createGeneralRouter(ctx: AppContext) {
