+++ src/tests/test-utils.ts
@@ -1,7 +1,7 @@
 import request from './hono-test-client.js';
 import { ctx, createTestUser } from './test-setup.js';
 import type { HonoTestAgent, HonoTestRequestChain } from './hono-test-client.js';
-import type { AppEnv } from '../http.js';
+import type { AppEnv } from '../type.js';
 import type { Hono } from 'hono';
 
 type Agent = HonoTestAgent;
