+++ src/type.ts
@@ -1,5 +1,20 @@
-import type { ErrorHandler, NotFoundHandler } from 'hono';
-import type { AppMiddleware, AppRequest, AppResponse } from './http.js';
+import type { Context, ErrorHandler, MiddlewareHandler, NotFoundHandler } from 'hono';
+import type { ContentfulStatusCode } from 'hono/utils/http-status';
+import type { Knex } from 'knex';
+import type { config } from './config.js';
+import type { CronService as CronServiceType } from './crons.js';
+import { createDatabase } from './db/db.js';
+import type { Libs } from './libs.js';
+import { createAssets } from './utils/assets.js';
+import { createAuth } from './utils/auth.js';
+import { createDate } from './utils/date.js';
+import { createHtml } from './utils/html.js';
+import { createMail } from './utils/mail.js';
+import { createNtfy } from './utils/ntfy.js';
+import { createRequest } from './utils/request.js';
+import { createSearch } from './utils/search.js';
+import { createUtil } from './utils/util.js';
+import { createValidation } from './utils/validation.js';
 
 export type DefaultSearchProviders = 'duckduckgo' | 'google' | 'yahoo' | 'bing';
 
@@ -14,6 +29,142 @@ export type ReminderType = 'once' | 'recurring';
 export type ReminderFrequency = 'daily' | 'weekly' | 'monthly';
 
 export type Env = 'production' | 'development' | 'testing';
+
+export interface AppSessionData {
+    redirectTo?: string | null;
+    user?: User | null;
+    input?: Record<string, unknown> | null;
+    errors?: Record<string, unknown> | null;
+    searchCount?: number;
+    cumulativeDelay?: number;
+    verifiedHiddenItems?: Record<string, number>;
+    hiddenItemsVerified?: boolean;
+    hiddenItemsVerifiedAt?: number;
+    userCachedAt?: number;
+    csrfToken?: string;
+    flash?: Record<string, string[]>;
+}
+
+export interface AppSession extends AppSessionData {
+    id: string;
+    save(callback?: (error?: Error) => void): void;
+    destroy(callback?: (error?: Error) => void): void;
+    regenerate(callback?: (error?: Error) => void): void;
+}
+
+export interface AppLocals {
+    state?: Record<string, any>;
+    utils?: Record<string, any>;
+    csrfToken?: string;
+}
+
+export interface AppRequest {
+    method: string;
+    url: string;
+    originalUrl: string;
+    path: string;
+    headers: Record<string, string>;
+    protocol: string;
+    hostname?: string;
+    query: Record<string, string>;
+    params: Record<string, string>;
+    body: any;
+    session: AppSession;
+    user: User | undefined;
+    logger: Logger;
+    ip?: string;
+    socket: { remoteAddress?: string };
+    get(name: string): string | undefined;
+    header(name: string): string | undefined;
+    flash(type: string, message?: string): string[];
+}
+
+export class AppResponse {
+    locals: AppLocals;
+    statusCode = 200;
+    response: Response | null = null;
+
+    private readonly headers = new Headers();
+
+    constructor(
+        private readonly c: AppContextContext,
+        private readonly ctx: AppContext,
+    ) {
+        this.locals = c.get('locals');
+    }
+
+    status(statusCode: number) {
+        this.statusCode = statusCode;
+        return this;
+    }
+
+    set(headers: Record<string, string>) {
+        for (const [key, value] of Object.entries(headers)) {
+            this.headers.set(key, value);
+        }
+        return this;
+    }
+
+    setHeader(name: string, value: string) {
+        this.headers.set(name, value);
+        return this;
+    }
+
+    json(data: unknown) {
+        this.applyHeaders();
+        this.response = this.c.json(data, this.statusCode as ContentfulStatusCode);
+        return this.response;
+    }
+
+    send(data: string | ArrayBuffer | Uint8Array) {
+        this.applyHeaders();
+        this.response = new Response(data as BodyInit, {
+            status: this.statusCode,
+            headers: this.headers,
+        });
+        return this.response;
+    }
+
+    render(view: string, options: Record<string, unknown> = {}) {
+        this.applyHeaders();
+        const html = this.ctx.utils.template.render(view, {
+            ...this.locals,
+            csrfToken: this.locals.csrfToken ?? '',
+            ...options,
+        });
+        this.response = this.c.html(html, this.statusCode as ContentfulStatusCode);
+        return this.response;
+    }
+
+    redirect(url: string) {
+        this.applyHeaders();
+        const statusCode = this.statusCode >= 300 && this.statusCode < 400 ? this.statusCode : 302;
+        this.response = this.c.redirect(url, statusCode as 301 | 302 | 303 | 307 | 308);
+        return this.response;
+    }
+
+    private applyHeaders() {
+        for (const [key, value] of this.headers.entries()) {
+            this.c.header(key, value);
+        }
+    }
+}
+
+export interface AppEnv {
+    Variables: {
+        body: any;
+        locals: AppLocals;
+        session: AppSession;
+        sessionChanged: boolean;
+        sessionDestroyed: boolean;
+        requestId: string;
+        user: User | undefined;
+        logger: Logger;
+    };
+}
+
+export type AppContextContext = Context<AppEnv>;
+export type AppMiddleware = MiddlewareHandler<AppEnv>;
 
 export type Bang = {
     /** Category of the bang (e.g., "Multimedia", "Online Services"). */
@@ -317,23 +468,6 @@ export type PaginateArrayOptions = {
     perPage: number;
     total: number;
 };
-
-import type { Knex } from 'knex';
-
-import { createDate } from './utils/date.js';
-import { createHtml } from './utils/html.js';
-import { createAuth } from './utils/auth.js';
-import { createMail } from './utils/mail.js';
-import { createNtfy } from './utils/ntfy.js';
-import { createUtil } from './utils/util.js';
-import { createSearch } from './utils/search.js';
-import { createRequest } from './utils/request.js';
-import { createValidation } from './utils/validation.js';
-import { createAssets } from './utils/assets.js';
-import type { CronService as CronServiceType } from './crons.js';
-import { createDatabase } from './db/db.js';
-import type { config } from './config.js';
-import type { Libs } from './libs.js';
 
 export type DateUtils = ReturnType<typeof createDate>;
 export type HtmlUtils = ReturnType<typeof createHtml>;
