+++ src/utils/util.ts
@@ -5,10 +5,10 @@ import type {
     ColumnPreferences,
     PaginateArrayOptions,
     TurnstileVerifyResponse,
+    AppRequest as Request,
 } from '../type.js';
 import http from 'node:http';
 import https from 'node:https';
-import type { AppRequest as Request } from '../http.js';
 
 const DEFAULT_SCREENSHOT_PREFETCH_BATCH_SIZE = 5;
 const DEFAULT_SCREENSHOT_PREFETCH_TIMEOUT_MS = 10000;
